This commit was generated by cvs2svn to compensate for changes in r13693, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@13694 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/crtc/crtc.c
@@ -698,7 +698,7 @@ void crtc_raster_draw_alarm_handler(CLOCK offset)
         int visible_height = crtc.current_line;
                         /* crtc.regs[6] * (crtc.regs[9] + 1); */
 
-        crtc.screen_yoffset = (crtc.screen_height - visible_height) / 2;
+        crtc.screen_yoffset = ((int)crtc.screen_height - visible_height) / 2;
         if (crtc.screen_yoffset < CRTC_SCREEN_BORDERHEIGHT)
             crtc.screen_yoffset = CRTC_SCREEN_BORDERHEIGHT;
